백준: 큰 수 (BIG) (240913)

+++ study/src/juyoungoh/beakjoon/_14928.java
@@ -10,7 +10,7 @@ public class _14928 {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         String num = br.readLine();
-        long remainder = 0;
+        int remainder = 0;
 
         for (int i = 0; i < num.length(); i++) {
             remainder = (remainder * 10 + (num.charAt(i) - '0')) % 20000303;
